Update Firefox Keyboard.ahk
MAJOR UPDATE: The script is now 570 lines instead of 4285 thats a reduction of over 85%!!! Plus it has extra features. If you press appskey + capslock it will toggle the appskey, so you no longer need to constantly hold the appskey if you want. and also the modifier keys are blocked from other applications, which means if you press a modifier the game youre playing no longer recieves that input!